requirements.txt
sentence_transformers
openai
retry
networkx
vllm
autoawq
llmcompressor